+++ resources/js/Components/Stats/RecentWorkoutsChart.vue
@@ -2,6 +2,7 @@
 import { jeton } from '@/Utils/couleurs'
 import { computed } from 'vue'
 import BaseChart from './BaseChart.vue'
+import { etiquetteDeDate } from '@/Utils/date'
 
 const props = defineProps({
     data: {
@@ -16,10 +17,7 @@ const donneesChronologiques = computed(() => [...props.data].reverse())
 const labels = computed(() =>
     donneesChronologiques.value.map((d) => {
         const date = new Date(d.started_at)
-        return date.toLocaleDateString('fr-FR', {
-            day: 'numeric',
-            month: 'short',
-        })
+        return etiquetteDeDate(date)
     }),
 )